+++ Sources/VCore/Deprecations.swift
@@ -7,7 +7,20 @@
 
 import SwiftUI
 
-// MARK: - Macros
+// MARK: - KeyPath Initializable Enumeration
+extension KeyPathInitializableEnumeration {
+    @available(*, deprecated, message: "Use initializer instead")
+    public static func aCase<Property>(
+        key keyPath: KeyPath<Self, Property>,
+        value: Property
+    ) -> Self?
+        where Property: Equatable
+    {
+        Self.allCases.first { $0[keyPath: keyPath] == value }
+    }
+}
+
+// MARK: - Uninitializable Macro
 @available(*, deprecated, renamed: "Uninitializable")
 @attached(member, names: named(init))
 public macro NonInitializable() = #externalMacro(